cards-main/requirements.txt
Flask
Flask-SQLAlchemy
requests
python-telegram-bot
Werkzeug
gunicorn
flask-cors
